ncurses 5.9 - patch 20120707
+ add ncurses_mingw.h to installed headers (prompted by patch by
  Juergen Pfeifer).
+ clarify return-codes from wgetch() in response to SIGWINCH (prompted
  by Novell #769788).
+ modify resizeterm() to always push a KEY_RESIZE onto the fifo, even
  if screensize is unchanged.  Modify _nc_update_screensize() to push a
  KEY_RESIZE if there was a SIGWINCH, even if it does not call
  resizeterm().  These changes eliminate the case where a SIGWINCH is
  received, but ERR returned from wgetch or wgetnstr because the screen
  dimensions did not change (Novell #769788).

+++ test/demo_termcap.c
@@ -29,7 +29,7 @@
 /*
  * Author: Thomas E. Dickey
  *
- * $Id: demo_termcap.c,v 1.21 2012/06/30 22:02:56 tom Exp $
+ * $Id: demo_termcap.c,v 1.22 2012/07/07 23:37:27 tom Exp $
  *
  * A simple demo of the termcap interface.
  */
@@ -89,7 +89,7 @@ static char *
 make_dbitem(char *p, char *q)
 {
     char *result = malloc(strlen(e_opt) + 2 + (size_t) (p - q));
-    sprintf(result, "%s=%.*s", e_opt, p - q, q);
+    sprintf(result, "%s=%.*s", e_opt, (int) (p - q), q);
     return result;
 }
